Make WatchedRealms Functional
